// src/enums/events.enum.ts
enum Events {
  UPDATE_ROOMS = "UPDATE_ROOMS",
  ROOM_ERROR_HANDLER = "ROOM_ERROR_HANDLER",
  NEW_ROOM = "NEW_ROOM",
  JOIN_ROOM = "JOIN_ROOM",
  JOIN_ROOM_DONE = "JOIN_ROOM_DONE",
  UPDATE_ROOM_PLAYERS_COUNT = "UPDATE_ROOM_PLAYERS_COUNT",
  LEAVE_ROOM = "LEAVE_ROOM",
  UPDATE_USERS_IN_ROOM = "UPDATE_USERS_IN_ROOM",
  UPDATE_USER_READY = "UPDATE_USER_READY",
  UPDATE_USER_READY_DONE = "UPDATE_USER_READY_DONE",
  UPDATE_USER_READY_CHANGE_BUTTON = "UPDATE_USER_READY_CHANGE_BUTTON",
  UPDATE_BEFORE_GAME_TIMER = "UPDATE_BEFORE_GAME_TIMER",
  START_GAME = "START_GAME",
  GAME_SESSION = "GAME_SESSION",
  UPDATE_GAME_TIMER = "UPDATE_GAME_TIMER",
  END_GAME = "END_GAME",
  SET_TEXT_COMPLETED = "SET_TEXT_COMPLETED",
  TEXT_UPDATE = "TEXT_UPDATE",
  TEXT_UPDATED_ROOM = "TEXT_UPDATED_ROOM",
  TEXT_UPDATED_CLIENT = "TEXT_UPDATED_CLIENT",
  SET_CLIENT_UNREADY = "SET_CLIENT_UNREADY",
  FORCE_UNAUTHORIZE = "FORCE_UNAUTHORIZE",
  SEND_RESULTS = "SEND_RESULTS",
  SET_CLEAR_TIMEOUT = 'SET_CLEAR_TIMEOUT',
  CLEAR_TIMEOUT = 'CLEAR_TIMEOUT'
}

export { Events };